utils/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dads5250"
version = "0.1.0"
description = "Shared utilities for DADS 5250: Generative AI in Practice"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.30",
    "tiktoken>=0.7",
    "ipywidgets>=8.0",
    "IPython",
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0"]

[tool.setuptools.packages.find]
include = ["dads5250*"]
